JC: Resolve issue with severe artifacts in image
This commit also includes several minor refactorings throughout the
codebase that I made while trying to debug the vertical lines artifact
bug.

The artifacts were caused by incorrectly setting the focus distance in
the camera constructor. The focus distance and aspect ratios have been
removed as options in the camera settings. The focus distance will be
computed automatically (it is the distance between look_at and
look_from) and the aspect ratio will be computed based on the width and
height parameters.

There is still a separate artifact issue when the camera aperture is set
to zero... this can be resolved by setting a small aperture such as
0.005. I don't know why this happens or what the fix is, so for now I
will add a TODO and just force the aperture to be at least 0.005.